refactor(web): move getServices to lib folder
